+++ src/FluentSerializer.Core/Changelog.md
@@ -5,6 +5,7 @@
 - [#146](https://github.com/Marvin-Brouwer/FluentSerializer/issues/146) Added `net7` support
 - [#227](https://github.com/Marvin-Brouwer/FluentSerializer/issues/227) Added `IParsable` converter
 - [#249](https://github.com/Marvin-Brouwer/FluentSerializer/issues/249) Added `IFormattable` converter
+- [#230](https://github.com/Marvin-Brouwer/FluentSerializer/issues/230) Improved iteration speed
 
 ## v3.0.0