+++ src/FluentSerializer.Core/Changelog.md
@@ -2,6 +2,10 @@
 
 ## @next
 
+-
+
+## v3.0.1
+
 - [#146](https://github.com/Marvin-Brouwer/FluentSerializer/issues/146) Added `net7` support
 - [#227](https://github.com/Marvin-Brouwer/FluentSerializer/issues/227) Added `IParsable` converter
 - [#249](https://github.com/Marvin-Brouwer/FluentSerializer/issues/249) Added `IFormattable` converter